pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "fitz-sage"
version = "0.11.0"
description = "Intelligent, honest knowledge retrieval in 5 minutes. No infrastructure. No boilerplate."
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Yan Fitzner" }
]
requires-python = ">=3.10"

# Keywords for PyPI search discoverability
keywords = [
    "rag",
    "retrieval",
    "augmented",
    "generation",
    "llm",
    "ai",
    "knowledge-base",
    "vector-database",
    "semantic-search",
    "document-processing",
    "embeddings",
    "nlp",
    "clara",
    "compression",
    "machine-learning",
]

# PyPI classifiers for better categorization
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
    "Topic :: Database",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.24",
    "typing-extensions>=4.7",
    "typer>=0.9",
    "jinja2>=3.1",
    # Cryptography for Fitz Cloud encryption (AES-256-GCM)
    "cryptography>=41.0",
    # PostgreSQL + pgvector for unified storage (vectors + tables)
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
    "pgvector>=0.2.0",
    "fitz-pgserver>=0.1.5",  # Fork of pgserver with Windows crash recovery fix
    # Lightweight document parsing (PDF, DOCX, PPTX — no ML)
    "pypdfium2>=4.0",
    "python-docx>=1.0",
    "python-pptx>=0.6",
    # Resilience for M2M auth (retry, circuit breaker)
    "tenacity>=8.0.0",
    "circuitbreaker>=2.1.3",
]

[project.optional-dependencies]
# Advanced document parsing (PDF layout analysis, table detection, figures)
docs = ["docling>=2.0"]

# Standalone code retrieval (no PostgreSQL, no pgvector, no docling)
code = []

# Local LLM (Ollama Python client)
local = ["ollama>=0.1.0"]

# REST API server
api = ["fastapi>=0.100.0", "uvicorn[standard]>=0.23.0"]

# Framework integrations
langchain = ["langchain-core>=0.2.0"]
llamaindex = ["llama-index-core>=0.10.0"]

# KRAG additional language support (tree-sitter extractors)
krag-typescript = ["tree-sitter>=0.22", "tree-sitter-typescript>=0.22"]
krag-java = ["tree-sitter>=0.22", "tree-sitter-java>=0.22"]
krag-go = ["tree-sitter>=0.22", "tree-sitter-go>=0.22"]
krag-all = [
    "tree-sitter>=0.22",
    "tree-sitter-typescript>=0.22",
    "tree-sitter-java>=0.22",
    "tree-sitter-go>=0.22",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "black>=23.0",
    "isort>=5.0",
    "ruff>=0.1.0",
    "psutil>=5.9",
    "hypothesis>=6.100.0",
]


[project.urls]
Homepage = "https://github.com/yafitzdev/fitz-sage"
Repository = "https://github.com/yafitzdev/fitz-sage"
Documentation = "https://github.com/yafitzdev/fitz-sage#readme"
Issues = "https://github.com/yafitzdev/fitz-sage/issues"
Changelog = "https://github.com/yafitzdev/fitz-sage/blob/main/CHANGELOG.md"
"Source Code" = "https://github.com/yafitzdev/fitz-sage"


[tool.setuptools.packages.find]
where = ["."]
include = ["fitz_sage*"]
exclude = ["tests*", "tools*", "examples*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.joblib"]

[project.scripts]
fitz = "fitz_sage.cli.cli:app"


[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"tools/contract_map/__main__.py" = ["E402"]
"tools/contract_map/common.py" = ["E402"]
# Test files: E402 is intentional - pytest markers must come before imports
"tests/**/*.py" = ["E402"]
# Examples: imports placed contextually for teaching purposes
"examples/**/*.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    # Tier markers (for CI pipeline control)
    "tier1: critical path tests - run on every commit (<30s)",
    "tier2: unit tests with mocks - run on PR (<2min)",
    "tier3: integration tests - run on merge to main (<10min)",
    "tier4: heavy tests - run nightly (30min+)",
    # Feature markers
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (real services)",
    "e2e: marks tests as end-to-end tests requiring real LLM and ingestion (fitz_krag)",
    "e2e_parser: marks tests as parser e2e tests (PDF/DOCX parsing)",
    "e2e_krag: marks tests as KRAG end-to-end tests requiring real LLM and ingestion",
    "e2e_krag_parser: marks tests as KRAG parser e2e tests (PDF/DOCX parsing)",
    "e2e_krag_formats: marks tests as KRAG format e2e tests (PPTX/XLSX/SQL/Go/Java/TS)",
    "postgres: marks tests as postgres-specific (pgvector, table store)",
    "llm: marks tests requiring real LLM API calls",
    "embeddings: marks tests requiring real embedding API calls",
    # Category markers
    "performance: marks tests as performance benchmarks",
    "scalability: marks tests as scalability tests (large corpus, concurrent)",
    "security: marks tests as security tests (injection, leakage, validation)",
    "chaos: marks tests as chaos/reliability tests (failure modes, recovery)",
    "property: marks tests as property-based tests (hypothesis)",
]

[tool.mutmut]
# Full mutation testing - run overnight locally or in weekly CI
paths_to_mutate = "fitz_sage/core/"
# Run unit + integration tests for thorough coverage
runner = "python -m pytest tests/unit/ tests/integration/ -x -q --tb=no --ignore=tests/unit/test_postgres_recovery.py"
# Test execution timeout (generous for full suite)
test_time_multiplier = 3.0
test_time_base = 10.0